+++ npm/vite-dev-server/src/startServer.ts
@@ -14,10 +14,17 @@ interface Options {
 }
 
 export interface StartDevServer {
-  /* this is the Cypress options object */
+  /**
+   * the Cypress options object
+   */
   options: Options
-  /* support passing a path to the user's webpack config */
-  viteConfig?: UserConfig // TODO: implement taking in the user's vite configuration. Right now we don't
+  /**
+   * By default, vite will use your vite.config file to
+   * Start the server. If you need additional plugins or
+   * to override some options, you can do so using this.
+   * @optional
+   */
+  viteConfig?: UserConfig
 }
 
 const resolveServerConfig = async ({ viteConfig, options }: StartDevServer): Promise<InlineConfig> => {
@@ -36,10 +43,14 @@ const resolveServerConfig = async ({ viteConfig, options }: StartDevServer): Pro
   // only cjs compiler-core accepts using prefixIdentifiers in slots which vue test utils use.
   // Could we resolve this usage in test-utils?
   finalConfig.resolve = finalConfig.resolve || {}
-  finalConfig.resolve.alias = {
-    ...finalConfig.resolve.alias,
-    '@vue/compiler-core': resolve(dirname(require.resolve('@vue/compiler-core')), 'dist', 'compiler-core.cjs.js'),
-  },
+  try {
+    finalConfig.resolve.alias = {
+      ...finalConfig.resolve.alias,
+      '@vue/compiler-core': resolve(dirname(require.resolve('@vue/compiler-core')), 'dist', 'compiler-core.cjs.js'),
+    }
+  } catch (e) {
+    // Vue 3 is not installed
+  }
 
   finalConfig.server = finalConfig.server || {}